Force API Key authentication for public endpoints
